move db and models to separate packages
