+++ backend/src/models/__init__.py
@@ -1,0 +1,2 @@
+from .recipient import Recipient
+from .campaign import Campaign
